basic changes

+++ JazzBot/Commands/MusicCommands.cs
@@ -131,7 +131,7 @@ namespace JazzBot.Commands
 				{
 					if (res >= 1 && res <= playNexts.Count)
 					{
-						var guild = await db.Guilds.SingleOrDefaultAsync(g => g.IdOfGuild == context.Guild.Id).ConfigureAwait(false);
+						var guild = await db.Guilds.SingleOrDefaultAsync(g => g.IdOfGuild ==(long) context.Guild.Id).ConfigureAwait(false);
 						db.Dispose();
 						this.GuildMusic.EnqueueToPlayNext(playNexts[res - 1].PathToFile);
 						await context.RespondAsync(embed: EmbedTemplates.ExecutedByEmbed(context.Member, context.Guild.CurrentMember)
