+++ JazzBot/Data/GuildMusicData.cs
@@ -96,7 +96,7 @@ namespace JazzBot.Data
 			this.Guild = guild;
 			this.Lavalink = lavalink;
 			var db = new DatabaseContext();
-			var dguild = db.Guilds.SingleOrDefault(x => x.IdOfGuild == this.Guild.Id);
+			var dguild = db.Guilds.SingleOrDefault(x => x.IdOfGuild == (long)this.Guild.Id);
 			db.Dispose();
 			this.Seed = dguild.Seed;
 			this.IdOfCurrentSong = dguild.IdOfCurrentSong;
@@ -140,9 +140,9 @@ namespace JazzBot.Data
 		private void UpdateDB()
 		{
 			var db = new DatabaseContext();
-			var guild = db.Guilds.SingleOrDefault(x => x.IdOfGuild == this.Guild.Id);
+			var guild = db.Guilds.SingleOrDefault(x => x.IdOfGuild == (long)this.Guild.Id);
 			guild.IdOfCurrentSong = this.IdOfCurrentSong;
-			guild.IdOfGuild = this.Guild.Id;
+			guild.IdOfGuild = (long)this.Guild.Id;
 			guild.PlaylistName = this.PlaylistName;
 			guild.Seed = this.Seed;
 			
@@ -270,7 +270,7 @@ namespace JazzBot.Data
 			var db = new DatabaseContext();
 			if (await db.Playlist.Select(x => x.PlaylistName).ContainsAsync(PlaylistName).ConfigureAwait(false))
 			{
-				var guild = await db.Guilds.SingleOrDefaultAsync(x => x.IdOfGuild == this.Guild.Id).ConfigureAwait(false);
+				var guild = await db.Guilds.SingleOrDefaultAsync(x => x.IdOfGuild == (long)this.Guild.Id).ConfigureAwait(false);
 				this.PlaylistName = playlistName;
 				guild.PlaylistName = playlistName;
 				this.Shuffle();
